Locking in the last commit
- Root namespace for test projects will be the same as the root
namespace as the project they're testing.
- Tests for Example.Namespace in the root project will reside in
Example.Namespace.Tests in the test project
- Test helpers will be in the root namespace in the test project, as
internal.

This is a bit awkward, but it's to avoid arbitrarily having to add using
statements in the test code for Example.Namepsace because the test code
happens to live in Example.Tests.Namespace.
And the Jenkins xUnit plugin seems to have a problem with putting the
tests in the same namespace as the SUT for some reason?