More work.
It took an awful lot of awful hacks to get here!  But "something's
happening".
- Extensible network message handlers have appeared.
- Temporarily created a Blockchain.info chain store to use while I work
on more packets.
--- This is so I can observe an "inv" message on the live network, then
send the appropriate "getdata" message so that when I get a "tx"
message, I can bootstrap an ITransaction from just a couple of
transactions' worth of data without having to write more and more
message handlers.
- Copy-paste and hacks galore.  This really needs to be trimmed down,
now that I've got something that "works".